refactor: update imports and data structures for consistency
- Changed import paths from `graph_typing` to `types` for better organization and clarity in the GraphQL-related files.
- Updated the `tasks` property in the `ProblemStatement` interface from an `IdMap<Task>` to a `Task[]` array for improved data handling and consistency across the codebase.
- Adjusted the `threadsService` to initialize `tasks` as an empty array instead of an object, aligning with the new structure.
